requirements/docsitalia.txt
-r docsitalia-base.txt
django-storages[azure]>=1.7.2
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
-r deploy.txt
